Extract functions from `maybeCombineDataset`
